src/requirements.txt
groq;
python-dotenv